backend/requirements.txt
Flask==2.2.2
requests==2.28.1
python-dotenv==0.21.0
psycopg2-binary==2.9.5  # PostgreSQL-Treiber
gunicorn==20.1.0
redis==4.3.4
rq==1.12.0
python-dateutil==2.8.2 #Für expires_at
PyJWT==2.8.0